requirements.txt
Django==5.2.10
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests>=2.32.3,<3
chardet<6
whitenoise
django-tailwind
django-allauth[socialaccount]
stripe